+++ src/common/net/Client.cpp
@@ -5,8 +5,8 @@
  * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
  * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
- * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
- * Copyright 2018 MoneroOcean      <https://github.com/MoneroOcean>, <support@moneroocean.stream>
+ * Copyright 2016-2019 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
+ * Copyright 2018-2019 MoneroOcean <https://github.com/MoneroOcean>, <support@moneroocean.stream>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -215,9 +215,11 @@ const char *Client::tlsVersion() const
 
 int64_t Client::submit(const JobResult &result)
 {
+#   ifndef XMRIG_PROXY_PROJECT
     if (result.clientId != m_rpcId) {
         return -1;
     }
+#   endif
 
     using namespace rapidjson;
 
@@ -340,7 +342,10 @@ bool Client::parseJob(const rapidjson::Value &params, int *code)
         return false;
     }
 
-    if (params.HasMember("variant")) {
+    if (params.HasMember("algo")) {
+        job.setAlgorithm(params["algo"].GetString());
+    }
+    else if (params.HasMember("variant")) {
         const rapidjson::Value &variant = params["variant"];
 
         if (variant.IsInt()) {
@@ -349,11 +354,6 @@ bool Client::parseJob(const rapidjson::Value &params, int *code)
         else if (variant.IsString()){
             job.setVariant(variant.GetString());
         }
-    }
-
-    // moved algo after variant parsing to override variant that is considered to be outdated now
-    if (params.HasMember("algo")) {
-        job.setAlgorithm(params["algo"].GetString());
     }
 
     if (!verifyAlgorithm(job.algorithm())) {